RC0075 - Integrate labelling with evaluation framework (#5)
We have the video labels stored in the DB, so now we can use them inside the evaluation framework instead of using csv files. They are currently used in generate_report.py

As part of this, we want to do a couple of QoL changes to the labelling tool and process. Ideally we want to download the video, label it and generate a report all from one place.

Additions:

load video from S3
load labels from DB
save labels to DB
generate report
upload report to S3