EVAL0042 - Update import button on exercise_video_annotator (#8)
Tasks

Make the labelling tool can import CSV files into the table.

Symptoms

When trying to import a CSV file into the table, the content in the CSV file is linked to the wrong column.

Problems

After a few updates it is no longer possible to import CSV files.

Solution

My method is similar to when the script imports the labels from the database. With my method I loaded the CSV file using pandas, iterated through each row and fill each column with the correct data.

Verification

N/A. No test for pavs.py. The script still works.